requirements.txt
pynput>=1.7.6
opencv-python>=4.5.0
mss>=6.1.0
numpy>=1.21.0
pywin32>=305
pyinstaller>=6.10.0
python-dotenv>=1.0.0
Pillow>=8.0.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
openai>=1.0.0